Add waiver options
